Almost finished the issues #21 and #23. Now we create a complete queue through the logical queue factory and this will : 1. Locate the host manager 2. Ask the host manager to create the queue and to choose a port (monitor + Xchanges) 3. Create the entry with the reference to the monitor in the global configuration
Still TODO:
A clean way to stop a monitor and an exchange. This could be done by JMX or directly by addressing a request socket on the monitor and exchange.

PS: the test on the create queue by script have been commented because we need to get the script and the jar installed.